parse/lex: cleanup, rename prec -> ops
